Re-combine NewsletterPreference classes and integrate nonce verification.
